Add Restful Support

+++ FEOP/FEOPServer/FEOPServerBlog/src/main/java/org/personal/mason/feop/server/blog/mvc/controllers/BlogController.java
@@ -1,7 +1,8 @@
 package org.personal.mason.feop.server.blog.mvc.controllers;
 
-import java.security.Principal;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.personal.mason.feop.server.blog.domain.model.Blog;
 import org.personal.mason.feop.server.blog.domain.model.BlogSetting;
@@ -11,7 +12,9 @@ import org.personal.mason.feop.server.blog.domain.service.BlogService;
 import org.personal.mason.feop.server.blog.domain.service.CategoryService;
 import org.personal.mason.feop.server.blog.domain.service.SeryService;
 import org.personal.mason.feop.server.blog.mvc.model.BlogModel;
-import org.personal.mason.feop.server.blog.mvc.utils.PrincipalUtils;
+import org.personal.mason.feop.server.blog.mvc.utils.AuthenticationUtils;
+import org.personal.mason.feop.server.blog.mvc.utils.ViewMapper;
+import org.personal.mason.feop.server.blog.utils.TimeUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -21,9 +24,6 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-
-import org.personal.mason.feop.server.blog.mvc.utils.ViewMapper;
-import org.personal.mason.feop.server.blog.utils.TimeUtils;
 
 @Controller
 public class BlogController {
@@ -48,7 +48,7 @@ public class BlogController {
 	}
 
 	/* Public Section of Blog operations */
-	@RequestMapping(value = "blog/view", method = RequestMethod.GET)
+	@RequestMapping(value = "/blog/view", method = RequestMethod.GET)
 	public String viewBlog(@RequestParam("id") Long id, Model model) {
 		Blog blog = blogService.findById(id);
 		BlogModel blogModel = BlogModel.revert(blog);
@@ -57,7 +57,7 @@ public class BlogController {
 		return ViewMapper.Blog_View.getViewName();
 	}
 
-	@RequestMapping(value = "blog/newest")
+	@RequestMapping(value = "/blog/newest")
 	public String getNewestBlogs(@RequestParam(value = "p", required = false) Integer page,
 			@RequestParam(value = "l", required = false) Integer size, Model model) {
 		if (page == null || page < 0) {
@@ -75,7 +75,7 @@ public class BlogController {
 		return ViewMapper.Index.getViewName();
 	}
 
-	@RequestMapping(value = "blog/search")
+	@RequestMapping(value = "/blog/search")
 	public String searchBlogs(@RequestParam("q") String query, @RequestParam(value = "p", required = false) Integer page,
 			@RequestParam(value = "l", required = false) Integer size, Model model) {
 		if (page == null || page < 0) {
@@ -93,7 +93,7 @@ public class BlogController {
 		return ViewMapper.Blog_List.getViewName();
 	}
 
-	@RequestMapping(value = { "blog", "blog/list" }, method = RequestMethod.GET)
+	@RequestMapping(value = { "/blog", "/blog/list" }, method = RequestMethod.GET)
 	public String findBlog(@RequestParam(value = "p", required = false) Integer page, @RequestParam(value = "l", required = false) Integer size,
 			@RequestParam(value = "c", required = false) Long categoryId, @RequestParam(value = "s", required = false) Long seryId, Model model) {
 		if (page == null || page < 0) {
@@ -121,15 +121,16 @@ public class BlogController {
 	}
 
 	/* Private Section of Blog Operations */
-	@RequestMapping(value = "m/blog/create", method = RequestMethod.GET)
+	@RequestMapping(value = "/my/blog/create", method = RequestMethod.GET)
 	public String createBlog(@ModelAttribute BlogModel blogModel) {
 		return ViewMapper.Blog_New.getViewName();
 	}
 
-	@RequestMapping(value = "m/blog/create", method = RequestMethod.POST)
-	public String saveBlog(@Validated BlogModel blogModel, BindingResult result, Principal principal) {
-		if (principal == null || principal.getName().isEmpty()) {
-			return "redirect:" + ViewMapper.Login.getViewName();
+	@RequestMapping(value = "/my/blog/create", method = RequestMethod.POST)
+	public String saveBlog(@Validated BlogModel blogModel, BindingResult result, HttpServletRequest request) {
+		String uid = AuthenticationUtils.getUid(request);
+		if (uid == null) {
+			return String.format("redirect:%s", "/user/login");
 		}
 
 		if (result.hasErrors()) {
@@ -140,8 +141,8 @@ public class BlogController {
 		blog.setCreateDate(TimeUtils.getCurrentTimestamp());
 		BlogModel.merge(blog, blogModel);
 		blog.setBlogSetting(new BlogSetting());
-		if (blogModel.getCategoryId() != null) {
-			Category category = categoryService.findById(blogModel.getCategoryId());
+		if (blogModel.getCategory() != null) {
+			Category category = categoryService.findById(blogModel.getCategory().getId());
 			blog.setCategory(category);
 		}
 
@@ -150,14 +151,14 @@ public class BlogController {
 			blog.setSery(sery);
 		}
 
-		blog.setAuthorName(PrincipalUtils.getUserName(principal));
-		blog.setAuthorUid(PrincipalUtils.getUid(principal));
+		blog.setAuthorName(AuthenticationUtils.getUserName(request));
+		blog.setAuthorUid(AuthenticationUtils.getUid(request));
 		blogService.save(blog);
 
 		return "redirect:/blog/view?id=" + blog.getId();
 	}
 
-	@RequestMapping(value = "m/blog/update", method = RequestMethod.GET)
+	@RequestMapping(value = "/my/blog/update", method = RequestMethod.GET)
 	public String updateBlog(@RequestParam("id") Long id, Model model) {
 		Blog blog = blogService.findById(id);
 
@@ -166,10 +167,11 @@ public class BlogController {
 		return ViewMapper.Blog_Edit.getViewName();
 	}
 
-	@RequestMapping(value = "m/blog/update", method = RequestMethod.PUT)
-	public String updateBlog(@Validated BlogModel blogModel, BindingResult result, Principal principal, Model model) {
-		if (principal == null || principal.getName().isEmpty()) {
-			return "redirect:/login";
+	@RequestMapping(value = "/my/blog/update", method = RequestMethod.PUT)
+	public String updateBlog(@Validated BlogModel blogModel, BindingResult result, HttpServletRequest request, Model model) {
+		String uid = AuthenticationUtils.getUid(request);
+		if (uid == null) {
+			return String.format("redirect:%s", "/user/login");
 		}
 
 		if (result.hasErrors()) {
@@ -178,21 +180,23 @@ public class BlogController {
 
 		Blog blog = new Blog();
 		BlogModel.merge(blog, blogModel);
-		blog.setAuthorName(PrincipalUtils.getUserName(principal));
-		blog.setAuthorUid(PrincipalUtils.getUid(principal));
+		blog.setAuthorName(AuthenticationUtils.getUserName(request));
+		blog.setAuthorUid(AuthenticationUtils.getUid(request));
 		Blog updatedBlog = blogService.update(blog);
 		BlogModel updatedBlogModel = BlogModel.revert(updatedBlog);
 		model.addAttribute("blog", updatedBlogModel);
 		return "redirect:/blog/view?id=" + blog.getId();
 	}
 
-	@RequestMapping(value = "m/blog/delete", method = RequestMethod.DELETE)
-	public void deleteBlog(@RequestParam("id") Long id) {
-		blogService.delete(id);
+	@RequestMapping(value = "/my/blog/delete", method = RequestMethod.DELETE)
+	public void deleteBlog(HttpServletRequest request, @RequestParam("id") Long id) {
+		String uid = AuthenticationUtils.getUid(request);
+		if (uid != null)
+			blogService.delete(id);
 	}
 
-	@RequestMapping(value = { "m/blog/list" }, method = RequestMethod.GET)
-	public String findMyBlog(Principal principal, @RequestParam(value = "p", required = false) Integer page,
+	@RequestMapping(value = { "/my/blog/list" }, method = RequestMethod.GET)
+	public String findMyBlog(HttpServletRequest request, @RequestParam(value = "p", required = false) Integer page,
 			@RequestParam(value = "l", required = false) Integer size, Model model) {
 		if (page == null || page < 0) {
 			page = 0;
@@ -202,7 +206,7 @@ public class BlogController {
 			size = 10;
 		}
 
-		String uid = PrincipalUtils.getUid(principal);
+		String uid = AuthenticationUtils.getUid(request);
 		List<Blog> blogs = blogService.findByAuthorUid(uid, page, size);
 		Long bcounts = blogService.getCount(uid);
 		model.addAttribute("blogs", blogs);
